Add testcases for DataCubes row-correctness
Four testcases are added to check the row-correctness of the DataCube class.
For each dimension (1D - 4D) a DataCube is filled with numbers (1, 2, ..., n).
After this initialization the inner vector is testest to represent this row-wise layout.

Additionally, a line is wrapped to ensure the Google C++ Style Guidline
and some includes are removed, since the respective header files are not used.